rewritten uncode_args, sample oerations add, ld
