+++ vm/Makefile
@@ -17,6 +17,7 @@ SOURCE = corewar_main.c \
         process_struct.c \
         processing.c \
         fetch.c \
+        uncode_args.c \
         operations/1_live.c \
         operations/2_ld.c \
         operations/3_st.c \
