+++ vm/fetch.c
@@ -56,47 +56,21 @@ int		check_cod_byte(int op_num, unsigned char codbyte, int *step, uint32_t *args
 	return (validate_cod_byte(op_num, args));
 }
 
-uint32_t	uncode_arg(unsigned char *field, size_t *curr, int size)
+void		fetch(t_general *gen, t_process *process, int op_num)
 {
-	size_t		j;
-	uint32_t	res;
-
-	j = 0;
-	res = 0;
-	while (j < size)
-	{
-		*curr = check_pc(*curr);
-		res = res << 8;
-		res |= field[(*curr)];
-		(*curr)++;
-		j++;
-	}
-	return (res);
-}
-
-void	fetch_args(unsigned char *field, t_process *proc, int op_num, uint32_t *args)
-{
-	size_t		i;
+	int 		step;
+	uint32_t	args[MAX_ARGS_NUMBER];
 	size_t		curr;
 
-	curr = check_pc(proc->pc + 1 + (op[op_num].coding_byte));
-	i = 0;
-	while (i < op[op_num].nbr_arg)
+	step = 0;
+	curr = check_pc(process->pc + 1);
+	while (step < MAX_ARGS_NUMBER)
+		args[step++] = 0;
+	step = 1;
+	if (check_cod_byte(op_num, gen->field[curr], &step, args))
 	{
-		if (args[i] == T_REG)
-		{
-			args[i] = (uint32_t)field[curr];
-			curr = check_pc(curr + 1);
-		}
-		else if (args[i] == T_DIR)
-			args[i] = uncode_arg(field, &curr, op[op_num].flag_direct_size);
-		else if (args[i] == T_IND)
-		{
-			args[i] = uncode_arg(field, &curr, IND_READ);
-
-			curr = proc->pc + args[i];
-			args[i] = uncode_arg(field, &curr, IND_SIZE);
-		}
-		i++;
+		op[op_num].f(gen, process, op_num, args);
 	}
+	if (op_num != 8) // not jump
+		process->pc += step;
 }
