+++ vm/operations/2_ld.c
@@ -1,12 +1,20 @@
 #include "../vm.h"
 
-void		ld_op(t_general *gen, t_process *process, int op_num, uint32_t *args)
+void	ld_op(t_general *gen, t_process *process, int op_num, uint32_t *args)
 {
-	if (args[1] >= 0 && args[1] <= 16)
+	uint32_t		args_val[MAX_ARGS_NUMBER];
+
+	args_copy(args, args_val, op[op_num].nbr_arg);
+	uncode_args(gen->field, process, op_num, args_val);
+	if (args_val[1] >= 1 && args_val[1] <= 16)
 	{
-		args[0] = args[0] % IDX_MOD;
-		process->reg[args[1]] = args[0];
-		process->carry = (process->reg[args[]] == 0) ? 1 : 0;
+		if (args[0] == T_DIR)
+			jfksldfjslkdfj;
+		else
+			jfksldfjslkdfj;
+//		args[0] = args[0] % IDX_MOD;
+//		process->reg[args[1]] = args[0];
+		process->carry = (process->reg[args_val[1]] == 0) ? 1 : 0;
 	}
 	return ;
 }
