+++ vm/operations/4_add.c
@@ -1,9 +1,9 @@
 
 #include "../vm.h"
 
-void		add_op(t_general *gen, t_process *process, int op_num, uint32_t *args)
+void	add_op(t_general *gen, t_process *process, int op_num, uint32_t *args)
 {
-	fetch_args(gen->field, process, op_num, args);
+	uncode_args(gen->field, process, op_num, args);
 	if (args[0] >= 1 && args[0] <= REG_NUMBER &&
 		args[1] >= 1 && args[1] <= REG_NUMBER &&
 		args[2] >= 1 && args[2] <= REG_NUMBER)
