+++ vm/processing.c
@@ -58,22 +58,3 @@ void		process(t_general *gen)
 		ptr = ptr->next;
 	}
 }
-
-void		fetch(t_general *gen, t_process *process, int op_num)
-{
-	int 		step;
-	uint32_t	args[MAX_ARGS_NUMBER];
-	size_t		curr;
-
-	step = 0;
-	curr = check_pc(process->pc + 1);
-	while (step < MAX_ARGS_NUMBER)
-		args[step++] = 0;
-	step = 1;
-	if (check_cod_byte(op_num, gen->field[curr], &step, args))
-	{
-//		fetch_args(gen->field, process, op_num, args);
-		op[op_num].f(gen, process, op_num, args);
-	}
-	process->pc += step;
-}
