+++ vm/vm.h
@@ -87,9 +87,12 @@ size_t				kill_process(t_process **head);
 
 size_t				check_pc(size_t pc);
 void				fetch(t_general *gen, t_process *process, int op_num);
-void				fetch_args(unsigned char *field, t_process *proc, int op_num, uint32_t *args);
 int					check_cod_byte(int op_num, unsigned char codbyte, int *step, uint32_t *args);
 int					validate_cod_byte(int op_num, uint32_t *args);
+
+void				uncode_args(unsigned char *field, t_process *proc, int op_num, uint32_t *args);
+uint32_t			convert_arg(unsigned char *field, size_t *curr, int size);
+void				args_copy(uint32_t *args, uint32_t *args_val, int nbr_arg);
 
 void				live_op(t_general *gen, t_process *process, int op_num, uint32_t *args);
 void				ld_op(t_general *gen, t_process *process, int op_num, uint32_t *args);
